+++ docs/props/CallbacksProps.tsx
@@ -1,5 +1,0 @@
-import { CallbacksProps as Props } from '../../src/types';
-
-export default function CallbacksProps(_props: Props): JSX.Element {
-  return null;
-}