+++ docs/props/OptionsProp.tsx
@@ -1,0 +1,5 @@
+import { OptionsProp as Props } from '../../src/types';
+
+export default function OptionsProp(_props: Props): JSX.Element {
+  return null;
+}