+++ docs/props/OptionsProps.tsx
@@ -1,5 +1,0 @@
-import { OptionsProps as Props } from '../../src/types';
-
-export default function OptionsProps(_props: Props): JSX.Element {
-  return null;
-}