+++ src/index.tsx
@@ -5,16 +5,7 @@ import seedrandom from 'seedrandom';
 
 import { useResponsiveSVGSelection } from './hooks';
 import render from './render';
-import {
-  Callbacks,
-  CallbacksProps,
-  MinMaxPair,
-  Options,
-  OptionsProps,
-  Scale,
-  Spiral,
-  Word,
-} from './types';
+import * as types from './types';
 import { getDefaultColors, getFontScale, getText, rotate } from './utils';
 
 const MAX_LAYOUT_ATTEMPTS = 10;
@@ -22,11 +13,11 @@ const SHRINK_FACTOR = 0.95;
 
 export * from './types';
 
-export const defaultCallbacks: Callbacks = {
-  getWordTooltip: ({ text, value }: Word): string => `${text} (${value})`,
+export const defaultCallbacks: types.Callbacks = {
+  getWordTooltip: ({ text, value }) => `${text} (${value})`,
 };
 
-export const defaultOptions: Options = {
+export const defaultOptions: types.Options = {
   colors: getDefaultColors(),
   deterministic: false,
   enableTooltip: true,
@@ -36,9 +27,8 @@ export const defaultOptions: Options = {
   fontWeight: 'normal',
   padding: 1,
   rotationAngles: [-90, 90],
-  rotations: undefined,
-  scale: Scale.Sqrt,
-  spiral: Spiral.Rectangular,
+  scale: types.Scale.Sqrt,
+  spiral: types.Spiral.Rectangular,
   transitionDuration: 600,
 };
 
@@ -46,7 +36,7 @@ export interface Props {
   /**
    * Callbacks to control various word properties and behaviors.
    */
-  callbacks?: CallbacksProps;
+  callbacks?: types.CallbacksProp;
   /**
    * Maximum number of words to display.
    */
@@ -54,19 +44,19 @@ export interface Props {
   /**
    * Set minimum [width, height] values for the SVG container.
    */
-  minSize?: MinMaxPair;
+  minSize?: types.MinMaxPair;
   /**
    * Configure wordcloud with various options.
    */
-  options?: OptionsProps;
+  options?: types.OptionsProp;
   /**
    * Set explicit [width, height] values for the SVG container.  This will disable responsive resizing.
    */
-  size?: MinMaxPair;
+  size?: types.MinMaxPair;
   /**
    * An array of word.  A word is an object that must contain the 'text' and 'value' keys.
    */
-  words: Word[];
+  words: types.Word[];
 }
 
 export default function Wordcloud({
@@ -76,16 +66,16 @@ export default function Wordcloud({
   options,
   size: initialSize,
   words,
-}: Props): React.ReactElement {
+}: Props): JSX.Element {
+  const mergedCallbacks = { ...defaultCallbacks, ...callbacks };
+  const mergedOptions = { ...defaultOptions, ...options };
+
   const [ref, selection, size] = useResponsiveSVGSelection(
     minSize,
     initialSize,
   );
 
-  useEffect((): void => {
-    const mergedCallbacks = { ...defaultCallbacks, ...callbacks };
-    const mergedOptions = { ...defaultOptions, ...options };
-
+  useEffect(() => {
     if (selection) {
       const {
         deterministic,
@@ -102,18 +92,16 @@ export default function Wordcloud({
 
       const sortedWords = words
         .concat()
-        .sort((x, y): number => descending(x.value, y.value))
+        .sort((x, y) => descending(x.value, y.value))
         .slice(0, maxWords);
 
-      const random: () => number = deterministic
-        ? seedrandom('deterministic')
-        : seedrandom();
+      const random = deterministic ? seedrandom('deterministic') : seedrandom();
 
       const layout = cloud()
         .size(size)
         .padding(padding)
         .words(sortedWords)
-        .rotate((): number => {
+        .rotate(() => {
           if (rotations === undefined) {
             // default rotation algorithm
             return (~~(random() * 6) - 3) * 30;
@@ -128,13 +116,13 @@ export default function Wordcloud({
         .fontStyle(fontStyle)
         .fontWeight(fontWeight);
 
-      const draw = (fontSizes: MinMaxPair, attempts = 1): void => {
+      const draw = (fontSizes: types.MinMaxPair, attempts = 1): void => {
         layout
-          .fontSize((word: Word): number => {
+          .fontSize((word: types.Word) => {
             const fontScale = getFontScale(sortedWords, fontSizes, scale);
             return fontScale(word.value);
           })
-          .on('end', (computedWords: Word[]): void => {
+          .on('end', (computedWords: types.Word[]) => {
             /** KNOWN ISSUE: https://github.com/jasondavies/d3-cloud/issues/36
              * Recursively layout and decrease font-sizes by a SHRINK_FACTOR.
              * Bail out with a warning message after MAX_LAYOUT_ATTEMPTS.
@@ -170,7 +158,7 @@ export default function Wordcloud({
 
       draw(fontSizes);
     }
-  }, [callbacks, maxWords, options, selection, size, words]);
+  }, [maxWords, mergedCallbacks, mergedOptions, selection, size, words]);
 
   return <div ref={ref} />;
 }